% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_registry.R
\name{get_profile}
\alias{get_profile}
\title{Retrieve a registered profile}
\usage{
get_profile(name)
}
\arguments{
\item{name}{Profile name.}
}
\value{
A \code{krt_profile} object.
}
\description{
Retrieve a registered profile
}
\examples{
get_profile("asap")$columns
}
