% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{krt_vocab}
\alias{krt_vocab}
\title{Access a KRT controlled vocabulary}
\usage{
krt_vocab(key)
}
\arguments{
\item{key}{Vocabulary name (e.g. "resource_type", "new_or_reuse", "status").}
}
\value{
A character vector of allowed values.
}
\description{
Access a KRT controlled vocabulary
}
\examples{
krt_vocab("new_or_reuse")
}
