% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{krt_vocabularies}
\alias{krt_vocabularies}
\alias{krt_resource_types}
\alias{krt_new_or_reuse}
\alias{krt_statuses}
\alias{krt_approval_types}
\alias{krt_roles}
\alias{krt_redaction_levels}
\title{KRT controlled vocabularies}
\usage{
krt_resource_types()

krt_new_or_reuse()

krt_statuses()

krt_approval_types()

krt_roles()

krt_redaction_levels()
}
\value{
A character vector of allowed values.
}
\description{
Convenience accessors for the controlled vocabularies used across the package.
}
\examples{
krt_resource_types()
krt_new_or_reuse()
}
