% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing.R
\name{krt_write_attribution}
\alias{krt_write_attribution}
\title{Write attribution to a file}
\usage{
krt_write_attribution(x, path)
}
\arguments{
\item{x}{A \link{krt_tbl} or profile name.}

\item{path}{Output file path.}
}
\value{
The path, invisibly.
}
\description{
Write attribution to a file
}
\examples{
f <- tempfile(fileext = ".md")
krt_write_attribution(krt_example, f)
}
