% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest.R
\name{list_suggest_sources}
\alias{list_suggest_sources}
\title{List autocomplete sources}
\usage{
list_suggest_sources()
}
\value{
A character vector of source names.
}
\description{
List autocomplete sources
}
\examples{
list_suggest_sources()
}
