% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json.R
\name{read_krt_yaml}
\alias{read_krt_yaml}
\title{Read a KRT from canonical YAML}
\usage{
read_krt_yaml(input)
}
\arguments{
\item{input}{A file path or a YAML string.}
}
\value{
A \link{krt_tbl}.
}
\description{
Read a KRT from canonical YAML
}
\examples{
k <- read_krt_yaml(write_krt_yaml(krt_example))
identical(length(k$resources), length(krt_example$resources))
}
