% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{rrid_type}
\alias{rrid_type}
\title{Resource type implied by an RRID}
\usage{
rrid_type(rrid)
}
\arguments{
\item{rrid}{An RRID string (with or without the \verb{RRID:} prefix).}
}
\value{
The implied resource type (character) or \code{NA} if the authority is
unknown.
}
\description{
Resource type implied by an RRID
}
\examples{
rrid_type("RRID:AB_390204")
rrid_type("CVCL_0063")
}
