% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krt_tbl.R
\name{summary.krt_tbl}
\alias{summary.krt_tbl}
\title{Summarize a KRT}
\usage{
\method{summary}{krt_tbl}(object, ...)
}
\arguments{
\item{object}{A \link{krt_tbl}.}

\item{...}{Ignored.}
}
\value{
A data frame with resource counts per resource type, including the
number of newly generated versus reused resources.
}
\description{
Summarize a KRT
}
\examples{
summary(krt_example)
}
